Allow user to rename without extension
Basically this limited the user from renaming a file with an extension to a file without extension. I was hoping to previously avoid file corruption with this, however knowing that this limits the user from actually renaming a file to something without an extension, I decided on removing this 'limitation'.