Refactor theme code - Major improvements
With this, performance has been noticeably improved. The way I used to do it was rather horrible. Creating 14 different files, and getting the RGB from them was a bad idea. This commit makes it so that 3DShell uses one single file to store all the colours, and one themeConfig file which points to a theme directory, and the path to the colour config.